requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Django=>4.2,<5.1
django-autoslug==1.9.9
django-cogwheels==0.3
django-filter==23.5
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.6
l18n==2021.3
openpyxl==3.1.2
pillow==10.2.0
pillow_heif==0.15.0
polib==1.2.0
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
typing_extensions==4.10.0
urllib3==2.2.1
wagtail==5.2.3
wagtail-localize==1.8.2
wagtail-modeladmin==2.0.0
wagtailmenus==4.0
webencodings==0.5.1
Willow==1.6.3
